moved a util method to shared file and added illustration when there is no time series twin in the list
