+++ src/Components/ADT3DSceneBuilder/Internal/Behaviors/Widgets/WidgetBuilders/DataHistoryWidgetBuilder/DataHistoryWidgetBuilder.tsx
@@ -21,7 +21,10 @@ import {
     DOCUMENTATION_LINKS
 } from '../../../../../../../Models/Constants/Constants';
 import { isValidADXClusterUrl } from '../../../../../../../Models/Services/Utils';
-import { getYAxisTypeOptions } from '../../../../../../../Models/SharedUtils/DataHistoryUtils';
+import {
+    getQuickTimeSpanKeyByValue,
+    getYAxisTypeOptions
+} from '../../../../../../../Models/SharedUtils/DataHistoryUtils';
 import {
     IADXTimeSeriesConnection,
     IDataHistoryAggregationType,
@@ -30,9 +33,7 @@ import {
 } from '../../../../../../../Models/Types/Generated/3DScenesConfiguration-v1.0.0';
 import { ADT3DScenePageContext } from '../../../../../../../Pages/ADT3DScenePage/ADT3DScenePage';
 import { ADXConnectionInformationLoadingState } from '../../../../../../../Pages/ADT3DScenePage/ADT3DScenePage.types';
-import QuickTimesDropdown, {
-    getQuickTimeSpanKeyByValue
-} from '../../../../../../QuickTimesDropdown/QuickTimesDropdown';
+import QuickTimesDropdown from '../../../../../../QuickTimesDropdown/QuickTimesDropdown';
 import TooltipCallout from '../../../../../../TooltipCallout/TooltipCallout';
 import { getActionButtonStyles } from '../../../../Shared/LeftPanel.styles';
 import { getWidgetFormStyles } from '../../WidgetForm/WidgetForm.styles';
