+++ src/Components/QuickTimesDropdown/QuickTimesDropdown.tsx
@@ -15,7 +15,6 @@ import {
 import { TFunction } from 'i18next';
 import { useTranslation } from 'react-i18next';
 import { QuickTimeSpans } from '../../Models/Constants/Constants';
-import { QuickTimeSpanKey } from '../../Models/Constants';
 
 const getClassNames = classNamesFunction<
     IQuickTimesDropdownStyleProps,
@@ -66,18 +65,6 @@ const getQuickTimeSpanOptions = (t: TFunction): Array<IDropdownOption> => {
         text: t(`quickTimesDropdown.options.${timeSpan}`),
         data: QuickTimeSpans[timeSpan]
     }));
-};
-
-/** Returns QuickTimeSpanKey from given millisecond */
-export const getQuickTimeSpanKeyByValue = (
-    millis: number
-): QuickTimeSpanKey => {
-    let key: QuickTimeSpanKey;
-    const idx = Object.values(QuickTimeSpans).indexOf(millis);
-    if (idx !== -1) {
-        key = Object.keys(QuickTimeSpans)[idx] as QuickTimeSpanKey;
-    }
-    return key;
 };
 
 export default styled<
